BallArithmetic

Documentation for BallArithmetic.

BallArithmetic.collatz_upper_bound_L2_norm

BallArithmetic.svd_bound_L2_norm

BallArithmetic.svd_bound_L2_norm_inverse

BallArithmetic.svd_bound_L2_resolvent

BallArithmetic.upper_abs

BallArithmetic.upper_bound_L1_norm

BallArithmetic.upper_bound_L2_norm

BallArithmetic.upper_bound_L_inf_norm

BallArithmetic.collatz_upper_bound_L2_norm — Method

collatz_upper_bound_L2_norm(A::BallMatrix; iterates=10)

Give a rigorous upper bound on the ℓ² norm of the matrix A by using the Collatz theorem.

We use Perron theory here: if for two matrices with B positive |A| < B we have ρ(A)<=ρ(B) by Wielandt's theorem Wielandt's theorem

The keyword argument iterates is used to establish how many times we are iterating the vector of ones before we use Collatz's estimate.

BallArithmetic.svd_bound_L2_norm — Method

svd_bound_L2_norm(A::BallMatrix{T})

Returns a rigorous upper bound on the ℓ²-norm of the ball matrix A using the rigorous enclosure for the singular values implemented in svd/svd.jl

BallArithmetic.svd_bound_L2_norm_inverse — Method

svd_bound_L2_norm_inverse(A::BallMatrix)

Returns a rigorous upper bound on the ℓ²-norm of the inverse of the ball matrix A using the rigorous enclosure for the singular values implemented in svd/svd.jl

BallArithmetic.svd_bound_L2_resolvent — Method

svd_bound_L2_resolvent(A::BallMatrix, lam::Ball)

Returns a rigorous upper bound on the ℓ²-norm of the resolvent of A at λ, i.e., ||(A-λ)^{-1}||_{ℓ²}

BallArithmetic.upper_abs — Method

upper_abs(A)

Return a floating point matrix B whose entries are bigger or equal (componentwise) any of the entries of A

BallArithmetic.upper_bound_L1_norm — Method

upper_bound_L1_norm(A::BallMatrix{T})

Returns a rigorous upper bound on the ℓ¹-norm of the ball matrix A

BallArithmetic.upper_bound_L2_norm — Method

upper_bound_L_inf_norm(A::BallMatrix{T})

Returns a rigorous upper bound on the ℓ²-norm of the ball matrix A using the best between the Collatz bound and the interpolation bound

BallArithmetic.upper_bound_L_inf_norm — Method

upper_bound_L_inf_norm(A::BallMatrix{T})

Returns a rigorous upper bound on the ℓ-∞-norm of the ball matrix A